# backend/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Vendora API",
        "description": "Vendora personal finance API. Use Bearer token authentication on protected endpoints.",
        "contact": {
            "email": "admin@vendora.com"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000",
            "description": "Local development server"
        }
    ],
    "paths": {
...
        "/api/investments": {
            "get": {
                "tags": [
                    "Investments"
                ],
                "summary": "List all investments",
                "operationId": "0219250acdc66727d7781c2bbe38dd17",
                "responses": {
                    "200": {
                        "description": "Array of investments"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Investments"
                ],
                "summary": "Create an investment",
                "operationId": "650ba82fccad5f7634cd8f4e7580abbc",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "name",
                                    "type",
                                    "initial_amount",
                                    "current_amount",
                                    "purchase_date"
                                ],
                                "properties": {
                                    "name": {
                                        "type": "string"
                                    },
                                    "type": {
                                        "type": "string"
                                    },
                                    "account": {
                                        "type": "string"
                                    },
                                    "initial_amount": {
                                        "type": "number"
                                    },
                                    "current_amount": {
                                        "type": "number"
                                    },
                                    "purchase_date": {
                                        "type": "string",
                                        "format": "date"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Investment created"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/investments/{investment}": {
            "get": {
                "tags": [
                    "Investments"
                ],
                "summary": "Get a single investment",
                "operationId": "3586fe45c408d5d968f69c30bafe7d9e",
                "parameters": [
                    {
                        "name": "investment",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Investment object"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "put": {
                "tags": [
                    "Investments"
                ],
                "summary": "Update an investment",
                "operationId": "87877e3690e0733a7dd5717bd3afdfb0",
                "parameters": [
                    {
                        "name": "investment",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "name": {
                                        "type": "string"
                                    },
                                    "type": {
                                        "type": "string"
                                    },
                                    "account": {
                                        "type": "string"
                                    },
                                    "initial_amount": {
                                        "type": "number"
                                    },
                                    "current_amount": {
                                        "type": "number"
                                    },
                                    "purchase_date": {
                                        "type": "string",
                                        "format": "date"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Investment updated"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Investments"
                ],
                "summary": "Delete an investment",
                "operationId": "0a4739ecd1c8f3534ab210f341739a5d",
                "parameters": [
                    {
                        "name": "investment",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Investment deleted"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "scheme": "bearer"
            }
        }
    },
...
}